+++ goondex/cmd/cli/go.mod
@@ -3,7 +3,6 @@ module goondex/cmd/cli
 go 1.15
 
 replace goondex/pkg/spider v0.0.0 => ../../pkg/spider
-
 replace goondex/pkg/engine v0.0.0 => ../../pkg/engine
 
 require (
