+++ goondex/pkg/engine/engine_test.go
@@ -4,18 +4,28 @@ import (
 	"testing"
 )
 
-func Test_Search(t *testing.T) {
-	eng := Engine{}
+type StubScanner struct {
+}
+
+func (st *StubScanner) Scan(string, int) (map[string]string, error) {
 	links := map[string]string{
 		"yandex.ru":  "Яндекс",
 		"google.com": "Google",
 	}
-	eng.Index(links)
+	return links, nil
+}
+
+func Test_Search(t *testing.T) {
+	eng := Engine{
+		scanner: &StubScanner{},
+	}
+	eng.Scan("example.com")
+	want := "Яндекс"
 	got := eng.Search("yandex")
 	if len(got) != 1 {
 		t.Fatalf("длина хэша не равна 1")
 	}
-	if got["yandex.ru"] != links["yandex.ru"] {
-		t.Fatalf("для ключа yandex.ru ждали %s, получили %s", links["yandex.ru"], got["yandex.ru"])
+	if got["yandex.ru"] != want {
+		t.Fatalf("для ключа yandex.ru ждали %s, получили %s", want, got["yandex.ru"])
 	}
 }
