+++ goondex/pkg/engine/go.mod
@@ -2,4 +2,9 @@ module goondex/pkg/engine
 
 go 1.15
 
-require golang.org/x/net v0.0.0-20201016165138-7b1cca2348c0 // indirect
+replace goondex/pkg/spider v0.0.0 => ../spider
+
+require (
+	golang.org/x/net v0.0.0-20201016165138-7b1cca2348c0 // indirect
+	goondex/pkg/spider v0.0.0
+)
