+++ goondex/pkg/engine/scanner.go
