+++ goondex/pkg/spider/spider.go
@@ -9,9 +9,15 @@ import (
 	"golang.org/x/net/html"
 )
 
+type Spider struct{}
+
+func New() *Spider {
+	return &Spider{}
+}
+
 // Scan осуществляет рекурсивный обход ссылок сайта, указанного в URL,
 // с учётом глубины перехода по ссылкам, переданной в depth.
-func Scan(url string, depth int) (data map[string]string, err error) {
+func (s *Spider) Scan(url string, depth int) (data map[string]string, err error) {
 	data = make(map[string]string)
 
 	err = parse(url, url, depth, data)
